Finished Meals controller

+++ app/database/migrations/2014_03_18_125913_core_tables.php
@@ -21,7 +21,7 @@ class CoreTables extends Migration {
         Schema::create('ingredients', function ($table) {
             $table->increments('id');
             $table->integer('meal_id')->unsigned();
-            $table->integer('size')->unsigned();
+            $table->float('size')->unsigned();
             $table->string('unit');
             $table->string('name');
